requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
httpx>=0.27.0
anthropic>=0.25.0
rich>=13.7.0
markdownify>=0.11.6
python-dotenv>=1.0.0
lxml>=5.2.0
